net_device: re-factor code so as to not be specific to lwIP
